+++ testdata/tree/well_tested/src/lib.rs
@@ -10,6 +10,7 @@
 
 mod inside_mod;
 mod item_mod;
+mod methods;
 mod nested_function;
 mod result;
 pub mod simple_fns;